Introduce an interface for external GC backends
