+++ src/options.h
@@ -82,6 +82,11 @@
 // GC_SMALL_PAGE allocates objects in 4k pages
 // #define GC_SMALL_PAGE
 
+// third party GC configuration options
+
+// MMTK_GC: support an alternative GC backend
+// #define MMTK_GC
+
 
 // method dispatch profiling --------------------------------------------------
